드라이브 마지막2

+++ src/main/java/com/backend/controller/ShareController.java
@@ -4,6 +4,9 @@ import com.backend.document.drive.ShareLink;
 import com.backend.dto.request.drive.RemoveDepartmentRequestDto;
 import com.backend.dto.request.drive.ShareLinkRequest;
 import com.backend.dto.request.drive.ShareRequestDto;
+import com.backend.dto.request.drive.SharedUser;
+import com.backend.dto.response.drive.FolderDto;
+import com.backend.service.FolderService;
 import com.backend.service.ShareService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +20,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 @RestController
 @RequiredArgsConstructor
@@ -26,10 +30,20 @@ public class ShareController {
 
     private final SimpMessagingTemplate messagingTemplate;
     private final ShareService shareService;
+    private final FolderService folderService;
 
     @PostMapping("/share/drive/{type}/{id}")
-    public ResponseEntity<?> shareDepartments(@RequestBody ShareRequestDto request, @PathVariable String type, @PathVariable String id) {
+    public ResponseEntity<?> shareDepartments(@RequestBody ShareRequestDto request, @PathVariable String type, @PathVariable String id, HttpServletRequest httpServletRequest) {
         log.info("Request received with type: {}, id: {}, body: {}", type, id, request);
+
+        String uid = (String) httpServletRequest.getAttribute("uid");
+        boolean permission =  ReadWritePermission(id, uid);
+        if(!permission) {
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body("공유 권한이 없습니다.");
+
+        }
 
         boolean result = shareService.shareUser(request, type, id);
         if(result) {
@@ -42,10 +56,16 @@ public class ShareController {
 
     @PostMapping("/share/link")
     public ResponseEntity<?> createShareLink(@RequestBody ShareLinkRequest sharedRequest, HttpServletRequest request) {
-        String permission = "읽기";
         String uid = (String) request.getAttribute("uid");
         log.info("Request received with id: {}, body: {}", sharedRequest.getId(),uid);
 
+        boolean permission =  ReadWritePermission(sharedRequest.getId(), uid);
+        if(!permission) {
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body("공유 권한이 없습니다.");
+
+        }
         if(sharedRequest.getOwnerId().equals(uid)){
             ShareLink shared = shareService.generateToken(sharedRequest.getId(),uid);
 
@@ -77,8 +97,16 @@ public class ShareController {
 
 
     @PostMapping("/drive/share/department")
-    public ResponseEntity<?> shareDepartment(@RequestParam String id ,@RequestParam String type,@RequestBody ShareRequestDto departments ) {
+    public ResponseEntity<?> shareDepartment(@RequestParam String id ,@RequestParam String type,@RequestBody ShareRequestDto departments,HttpServletRequest servletRequest ) {
         log.info("여기지금,"+type+"id"+id+"request"+departments);
+        String uid = (String) servletRequest.getAttribute("uid");
+        boolean permission =  ReadWritePermission(id, uid);
+        if(!permission) {
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body("공유 권한이 없습니다.");
+
+        }
 
         boolean result = shareService.sharedDepartment(departments,type,id);
 
@@ -94,7 +122,13 @@ public class ShareController {
     public ResponseEntity<?> removeDepartment(@RequestBody RemoveDepartmentRequestDto request, HttpServletRequest servletRequest) {
         log.info("여기여기여기!!!!"+request);
         String remover = (String) servletRequest.getAttribute("uid");
+        boolean permission =  ReadWritePermission(request.getId(), remover);
+        if(!permission) {
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body("공유 권한이 없습니다.");
 
+        }
         if(remover.equals(request.getOwnerId())) {
 
             shareService.deletedDepartment(request);
@@ -116,5 +150,29 @@ public class ShareController {
         Map<String, Object> response = shareService.invitationInvaild(invitationId,uid);
         log.info(" 공유결과!!!!"+response);
         return ResponseEntity.ok().body(response);
+    }
+
+
+    private boolean ReadWritePermission(String id, String uid) {
+
+        FolderDto folder = folderService.getParentFolder(id);
+        if(!folder.getOwnerId().equals(uid)){
+            List<SharedUser> list = folder.getSharedUsers()
+                    .stream()
+                    .filter(sharedUser -> sharedUser.getId().equals(uid))
+                    .collect(Collectors.toList());
+
+            // 첫 번째 일치하는 권한 반환
+            if (list.isEmpty()) {
+                return false;
+            }else{
+                String permission = list.get(0).getPermission();
+                if ("읽기".equals(permission) || "수정".equals(permission)) {
+                    return false;
+                }
+            }
+
+        }
+        return true;
     }
 }
